+++ OCR/app.yaml
@@ -1,2 +1,17 @@
-runtime: python39
-entrypoint: gunicorn -b :$PORT app:app
+runtime: python
+env: flex
+entrypoint: gunicorn -b :$PORT pythonAPI.analyzer:app
+
+runtime_config:
+    operating_system: "ubuntu22"
+    runtime_version: "3.12"
+
+    
+automatic_scaling:
+  min_num_instances: 1
+  max_num_instances: 3
+
+resources:
+  cpu: 1
+  memory_gb: 2
+  disk_size_gb: 10
