+++ OCR/pythonAPI/analyzer.py
@@ -2,6 +2,8 @@ import pytesseract
 from PIL import Image, ImageEnhance, ImageFilter
 from flask_cors import CORS
 from flask import Flask, request, jsonify
+import sys; print(sys.path)
+from .utils import cropImage, convert_circled_numerals_to_arabic
 import re
 import cv2
 from skimage.filters import threshold_local
@@ -10,18 +12,6 @@ from pathlib import Path
 import base64
 import io
 import json
-
-from utils import cropImage
-
-def convert_circled_numerals_to_arabic(text):
-    circled_numeral_mapping = {
-        '①': '1', '②': '2', '③': '3', '④': '4',
-        '⑤': '5', '⑥': '6', '⑦': '7', '⑧': '8',
-        '⑨': '9', '⑩': '10', '⑪': '11', '⑫': '12',
-        '⑬': '13', '⑭': '14', '⑮': '15', '⑯': '16',
-        '⑰': '17', '⑱': '18', '⑲': '19', '⑳': '20'
-    }
-    return ''.join(circled_numeral_mapping.get(char, char) for char in text)
 
 app = Flask('Tesseract-OCR-API')
 app.json.ensure_ascii = False
@@ -59,11 +49,11 @@ def upload_image():
         # OCR configuration for Japanese text
         config_japanese = '--oem 1 --psm 6'
         text_japanese = pytesseract.image_to_string(img_enhanced, lang='jpn', config=config_japanese)
-        print(text_japanese)
+        # print(text_japanese)
 
         # Convert circled numerals to numerals in the OCR output
         result_text = convert_circled_numerals_to_arabic(text_japanese)
-        print(result_text)
+        # print(result_text)
 
         # Get the data using regex
         product_regex = re.compile(r"\*?\s*(.*?)\s*(\d{13})[^\r\n]*?(\d+)")
